Inclass Projects/3- DL/requirements.txt
# Deep Learning Dependencies
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1

# Data Science and Visualization
numpy==1.23.5
matplotlib==3.6.2
seaborn==0.12.1
pandas==1.5.2
scikit-learn==1.1.3
plotly==5.12.0
kaleido==0.2.1

# Jupyter Environment
jupyter==1.0.0
notebook==6.5.2
ipykernel==6.17.1

# Image Processing
pillow==9.3.0

# API Development
fastapi==0.88.0
uvicorn[standard]==0.20.0
python-multipart==0.0.5
pydantic==1.10.4
aiofiles==22.1.0

# Utilities
tqdm==4.64.1
requests==2.28.1
python-dotenv==0.21.0

# Development and Monitoring
tensorboard==2.11.0

# Docker and SSH
docker==6.0.1
paramiko==2.12.0 